Implemented Authentication in ConnectionService

+++ connectorservice/src/main/java/com/kienast/connectorservice/controller/ConnectionController.java
@@ -20,6 +20,7 @@ import com.kienast.connectorservice.dto.ConnectionAdapter;
 import com.kienast.connectorservice.dto.ConnectionStatusAdapter;
 import com.kienast.connectorservice.dto.ConnectionStoreAdapter;
 import com.kienast.connectorservice.dto.ConnectionStoreStatusAdapter;
+import com.kienast.connectorservice.exception.NotAuthorizedException;
 import com.kienast.connectorservice.model.Connection;
 import com.kienast.connectorservice.model.ConnectionStatus;
 import com.kienast.connectorservice.model.ConnectionStore;
@@ -31,18 +32,23 @@ import com.kienast.connectorservice.rest.api.model.ConnectionCommandResponseMode
 import com.kienast.connectorservice.rest.api.model.ConnectionModel;
 import com.kienast.connectorservice.rest.api.model.ConnectionStatusModel;
 import com.kienast.connectorservice.rest.api.model.ConnectionStoreIdModel;
+import com.kienast.connectorservice.rest.api.model.ConnectionStoreIdRequestModel;
 import com.kienast.connectorservice.rest.api.model.ConnectionStoreModel;
 import com.kienast.connectorservice.rest.api.model.ConnectionStoreStatusModel;
 import com.kienast.connectorservice.rest.api.model.ConnectionStoreStringModel;
 import com.kienast.connectorservice.rest.api.model.DestroyConnectionRequestModel;
 import com.kienast.connectorservice.service.ConnectionService;
 import com.kienast.connectorservice.service.ConnectionStoreService;
+import com.kienast.connectorservice.service.TokenService;
 
 @RestController
 public class ConnectionController implements ConnectionApi, ConnstoreApi {
 
 	@Autowired
 	private ConnectionService connectionService;
+	
+	@Autowired
+	private TokenService tokenService;
 	
 	
 	@Autowired
@@ -58,11 +64,18 @@ public class ConnectionController implements ConnectionApi, ConnstoreApi {
 		return ResponseEntity.ok(response);
 	}
 
+
 	
 	@Override
 	public ResponseEntity<ConnectionStatusModel> createConnection(
-			@Valid ConnectionStoreIdModel connectionStoreIdModel) {
-		CreateConnectionCommand command = new CreateConnectionCommand(connectionStoreIdModel.getId());
+			@Valid ConnectionStoreIdRequestModel connectionStoreIdRequestModel) {
+		
+		
+		if (tokenService.validateToken(connectionStoreIdRequestModel.getToken()) == null) {
+			throw(new NotAuthorizedException(connectionStoreIdRequestModel.getToken()));
+		}
+		
+		CreateConnectionCommand command = new CreateConnectionCommand(connectionStoreIdRequestModel.getId());
 		ConnectionStatus connectionStatus = connectionService.createConnection(command);
 		ConnectionStatusModel response = new ConnectionStatusAdapter(connectionStatus).createJson();
 		return ResponseEntity.ok(response);
@@ -73,6 +86,12 @@ public class ConnectionController implements ConnectionApi, ConnstoreApi {
 	@Override
 	public ResponseEntity<ConnectionStatusModel> destroyConnection(
 			@Valid DestroyConnectionRequestModel destroyConnectionRequestModel) {
+		
+		if (tokenService.validateToken(destroyConnectionRequestModel.getToken()) == null) {
+			throw(new NotAuthorizedException(destroyConnectionRequestModel.getToken()));
+		}
+		
+		
 		DestroyConnectionCommand command = new DestroyConnectionCommand(destroyConnectionRequestModel.getSession());
 		ConnectionStatus connectionStatus = connectionService.destroyConnection(command);
 		ConnectionStatusModel response = new ConnectionStatusAdapter(connectionStatus).createJson();
@@ -84,6 +103,13 @@ public class ConnectionController implements ConnectionApi, ConnstoreApi {
 	@Override
 	public ResponseEntity<ConnectionCommandResponseModel> addCommand(
 			@Valid ConnectionCommandRequestModel connectionCommandRequestModel) {
+		
+		if (tokenService.validateToken(connectionCommandRequestModel.getToken()) == null) {
+			throw(new NotAuthorizedException(connectionCommandRequestModel.getToken()));
+		}
+		
+		
+		
 		ShellCommand command = new ShellCommand(connectionCommandRequestModel.getSession(),
 				connectionCommandRequestModel.getCommand());
 		
@@ -109,6 +135,13 @@ public class ConnectionController implements ConnectionApi, ConnstoreApi {
 	
 	@Override
 	public ResponseEntity<ConnectionStoreStatusModel> addConnectionStore(@Valid ConnectionStoreModel connectionStoreModel) {
+		
+		if (tokenService.validateToken(connectionStoreModel.getToken()) == null) {
+			throw(new NotAuthorizedException(connectionStoreModel.getToken()));
+		}
+		
+		
+		
 		CreateConnectionStoreCommand command = new CreateConnectionStoreCommand(connectionStoreModel.getHostname(),
 				connectionStoreModel.getPort(), connectionStoreModel.getUsername(),
 				connectionStoreModel.getPassword());
@@ -123,6 +156,12 @@ public class ConnectionController implements ConnectionApi, ConnstoreApi {
 	@Override
 	public ResponseEntity<ConnectionStoreStatusModel> deleteConnectionStore(
 			@Valid ConnectionStoreModel connectionStoreModel) {
+		
+		if (tokenService.validateToken(connectionStoreModel.getToken()) == null) {
+			throw(new NotAuthorizedException(connectionStoreModel.getToken()));
+		}
+		
+		
 		DestroyConnectionStoreCommand command = new DestroyConnectionStoreCommand(connectionStoreModel.getHostname(),
 				connectionStoreModel.getPort(), connectionStoreModel.getUsername(), connectionStoreModel.getPassword());
 		ConnectionStoreStatus connStoreStatus = connectionStoreService.deleteStoredConnection(command);
@@ -134,12 +173,21 @@ public class ConnectionController implements ConnectionApi, ConnstoreApi {
 	@Override
 	public ResponseEntity<ConnectionStoreStatusModel> updateConnectionStore(Integer storeId,
 			@Valid ConnectionStoreModel connectionStoreModel) {
+		
+		if (tokenService.validateToken(connectionStoreModel.getToken()) == null) {
+			throw(new NotAuthorizedException(connectionStoreModel.getToken()));
+		}
+		
+		
 		UpdateConnectionStoreCommand command = new UpdateConnectionStoreCommand(storeId, connectionStoreModel.getHostname(),
 				connectionStoreModel.getPort(), connectionStoreModel.getUsername(), connectionStoreModel.getPassword());
 		ConnectionStoreStatus connStoreStatus = connectionStoreService.updateStoredConnection(command);
 		ConnectionStoreStatusModel response = new ConnectionStoreStatusAdapter(connStoreStatus).createJson();
 		return ResponseEntity.ok(response);
 	}
+
+
+
